Reorganize engine file structure
